requirements.txt
fid_score==0.1.3
numba==0.57.0
numpy==1.24.3
Pillow==10.3.0
pqdm==0.2.0
scipy==1.13.0
SimpleITK==2.3.1
skimage==0.0
torch==2.3.0
